Updated messages

+++ Assets/Scripts/network/messages/StepMessage.cs
@@ -5,9 +5,9 @@ namespace Pandora.Network.Messages {
         /// <summary>Time passed in ms since last step </summary>
         public uint StepTimeMs { get; private set; }
         public List<Message> Commands { get; private set; }
-        public int? mana { get; private set; }
+        public float? mana { get; private set; }
 
-        public StepMessage(uint time, List<Message> commands, int? mana) {
+        public StepMessage(uint time, List<Message> commands, float? mana) {
             this.StepTimeMs = time;
             this.Commands = commands;
             this.mana = mana;
